add overloaded method for calcFeetAndInchesToCentimeters with one parameter

+++ src/methodOverloadingDemo/CalculateCentimeters.java
@@ -41,7 +41,7 @@ public class CalculateCentimeters {
     public static double calcFeetAndInchesToCentimeters(double inches2) {
         if (inches2 >= 0) {
             double feet2 = (int) inches2 / 12;
-            double remainingInches = inches2 % 12;
+            double remainingInches = (int) inches2 % 12;
             System.out.println(feet2);
             System.out.println(remainingInches);
             // calling main method in overloaded one
